python_template/requirements.txt
# Requisitos para o Sistema de Cabine Fotográfica
kivy==2.1.0
opencv-python==4.7.0.72
pillow==9.4.0
# pywin32==305  # Apenas para ambiente Windows
# kivy_garden.filebrowser==1.0.0  # Não utilizado no código atual